server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pymongo>=4.8.0
python-dotenv>=1.0.1
livekit-agents>=0.9.0
livekit>=0.9.0
pydantic>=2.7.0
pydantic-core>=2.18.0
httpx>=0.25.0
